2020.2: move around / fix background
