+++ lib/atom-browser-view-browser.js
@@ -2,8 +2,8 @@
 import { CompositeDisposable, Disposable } from 'atom'
 
 const fs = require('fs')
-const ViewNavbar = require('../views/Navbar')
-const ViewWebview = require('../views/Webview')
+const ViewNavbar = require('./views/Navbar')
+const ViewWebview = require('./views/Webview')
 
 import {
    CONFIG_RELOAD_DELAY,
@@ -30,7 +30,7 @@ export default class AtomBrowserViewBrowser {
       this.element = document.createElement('div')
       this.element.setAttribute('id', 'atombrowser-view')
       this.element.innerHTML += ViewNavbar()
-      this.element.innerHTML += ViewWebview()
+      this.element.innerHTML += ViewWebview({ showBackground })
 
       this.getTitle = () => 'Atom Browser'
       this.getURI = () => this.uri
@@ -40,7 +40,6 @@ export default class AtomBrowserViewBrowser {
       // wait for the html to be appended
       this.html = this.selectHtml()
       this.setListeners()
-      this.applySettingViewBackground(this.showBackground)
       this.setURL(this.initialUrl)
    }
