+++ lib/atom-browser-view-icon.js
@@ -1,6 +1,6 @@
 'use babel';
 
-const ViewIcon = require('../views/Icon')
+const ViewIcon = require('./views/Icon')
 
 export default class AtomBrowserViewIcon {
    constructor({ onClick }) {
