+++ lib/atom-browser.js
@@ -134,6 +134,7 @@ export default {
 
    changeConfigWebviewBackground() {
       const showBackground = atom.config.get(CONFIG_WEBVIEW_BACKGROUND)
+      this.showBackground = showBackground
       this.view && this.view.applySettingViewBackground(showBackground)
    },
