+++ lib/views/Icon.js
