+++ lib/views/Navbar.js
