+++ package.json
@@ -124,8 +124,5 @@
             }
          }
       }
-   },
-   "dependencies": {
-      "inspirational-quotes": "^1.0.8"
    }
 }
