+++ styles/atom-browser.less
@@ -86,7 +86,7 @@
          opacity: 0;
          animation: atombrowser-in-up 0.2s ease-in-out forwards;
          animation-delay: 0.1s;
-         text-align: left;
+         text-align: center;
       }
    }
 }
